+++ src/utility/database/mod.go
@@ -40,6 +40,25 @@ func SetLastLoginDate(uin int) error {
 	return err
 }
 
+func GetUserDetailsDataByUIN(uin int) (network.User, error) {
+
+	var user network.User
+
+	row, err := Query("SELECT * from userdetails WHERE UIN= ?", uin)
+
+	if err != nil {
+		logging.Error("Database/GetUserData", "Failed to get userdata: %s", err.Error())
+		return user, err
+	}
+
+	row.Next()
+	row.Scan(&user.UIN, &user.AvatarBlob, &user.AvatarImageType, &user.StatusMessage, &user.LastLogin, &user.SignupDate)
+	row.Close()
+
+	return user, err
+
+}
+
 func GetAccountDataByEmail(email string) (network.Account, error) {
 
 	var acc network.Account
